02-simple-rag-with-streamlit-fair-trade/requirements.txt
django-pyhub-rag
python-dotenv
streamlit
sqlite-vec
openai
anthropic
